Move in chunker with GELF
